requirements.txt
django-heroku==0.3.1
gunicorn==19.9.0
django-filter==2.0.0
djangorestframework==3.7.7
django-background-tasks==1.2.0
pytz==2018.9
psycopg2==2.7.6
whitenoise==3.3.1

beautifulsoup4==4.6.3
chromedriver-binary==76.0.3809.68.0
django-pandas==0.6.1
pandas==0.23.4
selenium==3.141.0
requests==2.5.1
lxml==4.2.5
